# Sawagger/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
    "/admin/active": {
      "put": {
        "description": "启用指定管理员（恢复被禁用的账户）",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "管理员管理"
        ],
        "summary": "启用管理员",
        "parameters": [
          {
            "description": "被启用用户名",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AdminService.BannedUserRequest"
            }
          },
          {
            "type": "string",
            "description": "Bearer token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.ActiveAdminUserSuccessResponse"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.ActiveAdminUserErrorResponse"
            }
          }
        }
      }
    },
    "/admin/ban": {
      "put": {
        "description": "禁用指定管理员",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "管理员管理"
        ],
        "summary": "禁用管理员",
        "parameters": [
          {
            "description": "被禁用用户名",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AdminService.BannedUserRequest"
            }
          },
          {
            "type": "string",
            "description": "Bearer token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.BannedAdminUserSuccessResponse"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.BannedAdminUserErrorResponse"
            }
          }
        }
      }
    },
    "/admin/change_pwd": {
      "put": {
        "description": "修改指定管理员密码",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "管理员管理"
        ],
        "summary": "修改管理员密码",
        "parameters": [
          {
            "description": "用户名和新密码",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AdminService.ChangeAdminPasswordRequest"
            }
          },
          {
            "type": "string",
            "description": "Bearer token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.ChangeAdminPasswordSuccessResponse"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.ChangeAdminPasswordErrorResponse"
            }
          }
        }
      }
    },
    "/admin/create": {
      "post": {
        "description": "创建新的管理员账户",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "管理员管理"
        ],
        "summary": "创建管理员",
        "parameters": [
          {
            "description": "用户名、密码、角色",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AdminService.AdminCreate"
            }
          },
          {
            "type": "string",
            "description": "Bearer token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.CreateAdminSuccessResponse"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.CreateAdminErrorResponse"
            }
          }
        }
      }
    },
    "/admin/delete": {
      "delete": {
        "description": "根据用户名删除管理员",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "管理员管理"
        ],
        "summary": "删除管理员",
        "parameters": [
          {
            "description": "被删除用户名",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AdminService.DeletedUserRequest"
            }
          },
          {
            "type": "string",
            "description": "Bearer token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.DeleteAdminSuccessResponse"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.DeleteAdminErrorResponse"
            }
          }
        }
      }
    },
    "/admin/list": {
      "post": {
        "description": "获取分页管理员列表，可过滤用户名和状态",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "管理员管理"
        ],
        "summary": "获取管理员列表",
        "parameters": [
          {
            "type": "integer",
            "description": "可选，删除状态筛选",
            "name": "deleted",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "csv",
            "description": "可选，指定ID列表筛选",
            "name": "ids",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "当前页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "每页数量",
            "name": "page_size",
            "in": "query"
          },
          {
            "type": "string",
            "description": "可选，角色筛选",
            "name": "role",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "可选，状态筛选",
            "name": "status",
            "in": "query"
          },
          {
            "type": "string",
            "description": "可选，用户名模糊筛选",
            "name": "username",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Bearer token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.AdminListSuccessResponse"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/swaggerResponse.ErrorResponseDoc"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.AdminListErrorResponse"
            }
          }
        }
      }
    },
    "/admin/login": {
      "post": {
        "description": "使用用户名密码登录，返回 JWT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "管理员管理"
        ],
        "summary": "管理员登录",
        "parameters": [
          {
            "description": "用户名和密码",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AdminService.AdminLoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.AdminLoginSuccessResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.AdminLoginErrorResponse"
            }
          }
        }
      }
    },
    "/admin/logout": {
      "post": {
        "description": "登出当前管理员",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "管理员管理"
        ],
        "summary": "管理员登出",
        "parameters": [
          {
            "type": "string",
            "description": "Bearer token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.LogOutSuccessResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.LogOutErrorResponse"
            }
          }
        }
      }
    },
    "/admin/remember": {
      "post": {
        "description": "登录后返回访问令牌和刷新令牌",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "管理员管理"
        ],
        "summary": "管理员记住登录",
        "parameters": [
          {
            "description": "登录凭证",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AdminService.LoginCredentialsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.AdminRememberSuccessResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.AdminRememberErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "AdminService.AdminCreate": {
      "type": "object",
      "properties": {
        "deleted": {
          "type": "integer"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "status": {
          "type": "integer"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "AdminService.AdminLoginRequest": {
      "type": "object",
      "required": [
        "auth_status",
        "device_id",
        "ip",
        "password",
        "username"
      ],
      "properties": {
        "auth_status": {
          "description": "记住登录状态：1-记住 2-不记住",
          "type": "integer"
        },
        "device_id": {
          "description": "设备唯一标识，用于多设备管理",
          "type": "string"
        },
        "ip": {
          "description": "当前登录IP地址",
          "type": "string"
        },
        "password": {
          "description": "登录密码，必填",
          "type": "string"
        },
        "username": {
          "description": "登录用户名，必填",
          "type": "string"
        }
      }
    },
    "AdminService.BannedUserRequest": {
      "type": "object",
      "required": [
        "username"
      ],
      "properties": {
        "username": {
          "description": "待禁用的管理员用户名",
          "type": "string"
        }
      }
    },
    "AdminService.ChangeAdminPasswordRequest": {
      "type": "object",
      "required": [
        "new_password",
        "username"
      ],
      "properties": {
        "new_password": {
          "description": "新密码",
          "type": "string"
        },
        "username": {
          "description": "管理员用户名",
          "type": "string"
        }
      }
    },
    "AdminService.DeletedUserRequest": {
      "type": "object",
      "required": [
        "username"
      ],
      "properties": {
        "username": {
          "description": "待删除的管理员用户名",
          "type": "string"
        }
      }
    },
    "AdminService.LoginCredentialsRequest": {
      "type": "object",
      "properties": {
        "auth_status": {
          "description": "记住登录状态：1-记住 2-不记住",
          "type": "integer"
        },
        "device_id": {
          "description": "设备唯一标识，用于多设备会话管理",
          "type": "string"
        },
        "ip": {
          "description": "登录IP地址，用于安全审计",
          "type": "string"
        },
        "password": {
          "description": "登录明文密码，由后端加密验证",
          "type": "string"
        },
        "rt": {
          "description": "刷新令牌(Refresh Token)，用于获取新访问令牌",
          "type": "string"
        },
        "username": {
          "description": "登录用户名，唯一标识",
          "type": "string"
        }
      }
    },
...
    "adminSwaggerResponse.ActiveAdminUserErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "err msg"
        }
      }
    },
    "adminSwaggerResponse.ActiveAdminUserSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "管理员 xxx 已启用"
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "adminSwaggerResponse.AdminListErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "err msg"
        }
      }
    },
    "adminSwaggerResponse.AdminListSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "{page:1, page_size: 10, total_pages: 100,data: [id1_Detail, id2_Detail, id3_Detail]}"
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "adminSwaggerResponse.AdminLoginErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "密码错误"
        }
      }
    },
    "adminSwaggerResponse.AdminLoginSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1N..."
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "adminSwaggerResponse.AdminRememberErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "err_msg"
        }
      }
    },
    "adminSwaggerResponse.AdminRememberSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "{token:eyJhbGciOiJIUzI1N...,rt:eyJhbGciOiJIUzI1N...}"
        },
        "message": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "adminSwaggerResponse.BannedAdminUserErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "err msg"
        }
      }
    },
    "adminSwaggerResponse.BannedAdminUserSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "管理员 xxx 已被禁用"
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "adminSwaggerResponse.ChangeAdminPasswordErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "err msg"
        }
      }
    },
    "adminSwaggerResponse.ChangeAdminPasswordSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "管理员 xxx 密码已更新"
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "adminSwaggerResponse.CreateAdminErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "管理员 xxx 已存在"
        }
      }
    },
    "adminSwaggerResponse.CreateAdminSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "管理员 xxx 已创建"
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "adminSwaggerResponse.DeleteAdminErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "err msg"
        }
      }
    },
    "adminSwaggerResponse.DeleteAdminSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "管理员 xxx 已删除"
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "adminSwaggerResponse.LogOutErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "err msg"
        }
      }
    },
    "adminSwaggerResponse.LogOutSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "登出成功"
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
...
    "swaggerResponse.ErrorResponseDoc": {
      "type": "object",
      "properties": {
        "code": {
          "description": "状态码",
          "type": "integer",
          "example": -1
        },
        "data": {
          "description": "数据",
          "type": "string",
          "example": "null"
        },
        "msg": {
          "description": "信息",
          "type": "string",
          "example": "参数错误"
        }
      }
    },
...
  }
}